The normalized counts and test results of a STCompR object.

The normalized counts and test results of a STCompR object.

# S4 method for class 'STCompR'
stat_test(
  object,
  transform = c("None", "log2", "log10", "log"),
  count_only = FALSE,
  melted_count = FALSE,
  normalized = TRUE,
  features = NULL
)

Arguments

object

The STCompR object

transform

Whether the count are transformed (the pseudo count defined for the object is added).

count_only

Returns only counts.

melted_count

Returns a melted table when count_only is TRUE.

normalized

Counts are normalized.

features

The features that should be returned. Default to all (NULL).

Examples

example_dataset("11284296/files/cmp_xen")
stat_test(cmp_xen)
#>            xen_r1    xen_r2    log2_ratio      p_values          padj
#> Ano1     518.5353 1426.8653  1.458566e+00 2.301534e-171 5.370247e-171
#> Chat     846.8914  188.4539 -2.162033e+00  2.955175e-65  4.137245e-65
#> Ebf3     362.9004  821.1206  1.175806e+00  1.166087e-77  2.040652e-77
#> Kctd8   1303.0242 1303.0242 -1.761885e-15  3.070874e-14  3.070874e-14
#> Necab2  6084.2463 8238.1280  4.371760e-01 9.999889e-321 3.499961e-320
#> Nrp2    3877.8713 2398.7490 -6.927533e-01  8.626948e-16  1.006477e-15
#> Nwd2   10829.4384 3475.6285 -1.639331e+00 9.999889e-321 3.499961e-320