Spatial transcriptomic comparison class (STCompR).

This class represents the result of a spatial transcriptomic comparison between two STGrid objects. It stores information such as conditions, methods, raw counts, scaling factors, and statistical test results.

Slots

conditions

A character vector specifying the conditions being compared.

method

A character vector specifying the technology used for each condition.

raw_counts

A data.frame storing raw count data for each condition.

norm_counts

A data.frame storing normalized counts for each condition.

scaling_factor

A numeric vector representing the scaling factors used during normalization (to be divided).

stat_test

A data frame containing statistical test results, including log2 ratios, odds ratios, and p-values for each feature.

pseudo_count

The value for the pseudo_count.

neighborhood

a list of feature-feature neighborhood for the two conditions to be compared (not functional at the moment).

neighborhood_changes

A feature-feature matrix indicative of neighborhood changes (not functional at the moment).

Examples

# Example usage:
example_dataset("11284296/files/cmp_xen")
#> |-- INFO :  Creating a path for dataset installation: /home/runner/.STarlight/datasets 
#> |-- INFO :  Dataset 11284296/files/cmp_xen has been loaded. 
cmp_xen
#> |-- INFO :  An object of class STCompR 
#> |-- INFO :  xen_r1 Condition_1 
#> |-- INFO :  xen_r2 Condition_2 
#> |-- INFO :  Number of features:  7 
#> |-- INFO :  >>> Please, use show_st_methods(class = 'STCompR') to show availables methods <<<