Spatial transcriptomic analysis class (STGrid).

A class for representing spatial transcriptomic analysis results as x/y coordinates. This class stores transcript molecules or cell coordinates as a grid.

Slots

coord

A data frames representing spatial coordinates (empty by default).

bin_mat

A data frames representing the matrix (empty by default).

y_max

Numeric value representing the maximum y-coordinate (default is 0).

y_min

Numeric value representing the minimum y-coordinate (default is 0).

x_min

Numeric value representing the minimum x-coordinate (default is 0).

x_max

Numeric value representing the maximum x-coordinate (default is 0).

path

Character value representing the file path to the original data.

method

The technology used.

meta

A list containing meta-information about bins.

bin_size

Numeric value representing the bin size.

bin_x

Character vector containing the names of bins/windows along the x-axis.

bin_y

Character vector containing the names of bins/windows along the y-axis.

ripley_k_function

The result of Ripley's K function computation.

control

A regular expression indicating the string motif related to controls.

Examples

example_dataset()
#> |-- INFO :  Dataset 11284233/files/Xenium_Mouse_Brain_Coronal_7g has been loaded. 
Xenium_Mouse_Brain_Coronal_7g
#> |-- INFO :  An object of class STGrid 
#> |-- INFO :  Number of counts:  152958 
#> |-- INFO :  Number of features:  7 
#> |-- INFO :  Bin size:  25 
#> |-- INFO :  Number of bins (x axis):  217 
#> |-- INFO :  Number of bins (y axis):  141 
#> |-- INFO :  x_min:  30.848026 
#> |-- INFO :  x_max:  5446.949 
#> |-- INFO :  y_min:  21.983248 
#> |-- INFO :  y_max:  3534.117 
#> |-- INFO :  >>> Please, use show_st_methods() to show availables methods <<< 
show_st_methods(class = "STGrid")
#>  [1] "$"                    "$<-"                  "["                   
#>  [4] "[["                   "[[<-"                 "as_matrix"           
#>  [7] "bin_mat"              "bin_size"             "bin_x"               
#> [10] "bin_y"                "cc_neighborhood"      "check_features_exist"
#> [13] "col_names"            "compute_k_ripley"     "compute_module_score"
#> [16] "connected_components" "coord"                "create_hull"         
#> [19] "dim"                  "feat_names"           "gene_contrast"       
#> [22] "get_coord"            "hc_tree"              "meta"                
#> [25] "meta_names"           "multi_gene_contrast"  "nb_feat"             
#> [28] "nb_items"             "nbin_x"               "nbin_y"              
#> [31] "order_feat_by_ripley" "plot_rip_k"           "re_bin"              
#> [34] "ripley_k_function"    "rm_controls"          "row_names"           
#> [37] "show"                 "spatial_image"        "spatial_plot"        
#> [40] "summary"              "tab"                  "table_st"            
#> [43] "write_coord"