The normalized counts and test results of a STCompR object.

The normalized counts and test results of a STCompR object.

# S4 method for STCompR
stat_test(
  object,
  transform = c("None", "log2", "log10", "log"),
  count_only = FALSE,
  melted_count = FALSE,
  normalized = TRUE,
  features = NULL
)

Arguments

object

The STCompR object

transform

Whether the count are transformed (the pseudo count defined for the object is added).

count_only

Returns only counts.

melted_count

Returns a melted table when count_only is TRUE.

normalized

Counts are normalized.

features

The features that should be returned. Default to all (NULL).

Examples

example_dataset("11284296/files/cmp_xen")
stat_test(cmp_xen)
#>            xen_r1    xen_r2    log2_ratio      p_values          padj
#> Ano1     529.1529 1390.7928  1.392464e+00 3.946149e-162 9.207680e-162
#> Chat     851.5704  187.5828 -2.176621e+00  1.502379e-64  2.103330e-64
#> Ebf3     373.1685  822.6848  1.138404e+00  5.870865e-77  1.027401e-76
#> Kctd8   1331.8382 1331.8382 -4.805140e-16  1.370950e-15  1.370950e-15
#> Necab2  6315.5029 8237.5667  3.832673e-01 9.999889e-321 3.499961e-320
#> Nrp2    3969.7660 2425.1782 -7.107318e-01  4.365080e-16  5.092594e-16
#> Nwd2   11046.5325 3681.9833 -1.584777e+00 9.999889e-321 3.499961e-320